+++ src/html/about.html
@@ -104,77 +104,7 @@
           <h2 class="git__title">История коммитов из GitHub</h2>
           <a class="git__open" href="https://github.com/Maxim-Ba/Diploma-project/commits/master" target="_blank">Открыть в Github ></a>
         </header>
-        <div class="git__slider main-carousel" data-flickity='{ "wrapAround": true, "percentPosition": false }'>
-          <article class="card carousel-cell">
-            <h2 class="card__title">Title</h2>
-            <a href="#" class="card__link"  target="_blank">
-              <p class="card__date">14 августа, 2019</p>
-              <div class="card__container">
-                <img class="card__img" src="<%=require('../images/avatar.jpg')%>" alt="author`s avatar picture">
-                <div class="container-flex">
-                  <cite class="card__author">Антон Долинин</cite>
-                  <p class="card__email">ingaiva@yandex.ru</p>
-                </div>
-              </div>
-              <blockquote class="card__comment">
-                Emmet (formerly Zen Coding) is a web- developer’s toolkit that can
-                greatly improve your HTML & CSS workflow.
-              </blockquote>
-            </a>
-          </article>
-          <article class="card carousel-cell">
-            <h2 class="card__title">Title</h2>
-            <a href="#" class="card__link" target="_blank">
-              <p class="card__date">14 августа, 2019</p>
-              <div class="card__container">
-                <img class="card__img" src="<%=require('../images/avatar.jpg')%>" alt="author`s avatar picture">
-                <div class="container-flex">
-                  <cite class="card__author">Антон Долинин</cite>
-                  <p class="card__email">ingaiva@yandex.ru</p>
-                </div>
-              </div>
-              <blockquote class="card__comment">
-                Emmet (formerly Zen Coding) is a web- developer’s toolkit that can
-                greatly improve your HTML & CSS workflow.
-              </blockquote>
-            </a>
-          </article>
-          <article class="card carousel-cell">
-            <h2 class="card__title">Title</h2>
-            <a href="#" class="card__link" target="_blank">
-              <p class="card__date">14 августа, 2019</p>
-              <div class="card__container">
-                <img class="card__img" src="<%=require('../images/avatar.jpg')%>" alt="author`s avatar picture">
-                <div class="container-flex">
-                  <cite class="card__author">Антон Долинин</cite>
-                  <p class="card__email">ingaiva@yandex.ru</p>
-                </div>
-              </div>
-              <blockquote class="card__comment">
-                You can install this plugin directly from NetBeans Plugin Portal: in
-                NetBeans, open <span class="card__mark">Tools > Plugins > Available plugins</span> and find and install
-                Emmet plugin.
-              </blockquote>
-            </a>
-          </article>
-          <article class="card carousel-cell">
-            <h2 class="card__title">Title</h2>
-            <a href="#" class="card__link" target="_blank">
-              <p class="card__date">14 августа, 2019</p>
-              <div class="card__container">
-                <img class="card__img" src="<%=require('../images/avatar.jpg')%>" alt="author`s avatar picture">
-                <div class="container-flex">
-                  <cite class="card__author">Антон Долинин</cite>
-                  <p class="card__email">ingaiva@yandex.ru</p>
-                </div>
-              </div>
-              <blockquote class="card__comment">
-                With Emmet, you can type CSS-like expressions that can be dynamically
-                parsed, and
-                produce output depending on what you type in the abb... 
-              </blockquote>
-            </a>
-          </article>
+        <div class="git__slider">
         </div>
         <a class="git__button_mobile" href="https://github.com/Maxim-Ba/Diploma-project/commits/master" target="_blank">Открыть в Github</a>
       </section>
